Refactor filter DTOs into Filters.kt
Remove the separate Dto.kt and consolidate filter-related classes into Filters.kt. Rename Genre to FilterOption and add @Serializable annotations. Update Filter.Select types to use FilterOption and adjust query parameter use to read .value. Simplify building FilterList by using the deserialized FilterData arrays directly.
