--- src/vi/tranh18/src/eu/kanade/tachiyomi/extension/vi/tranh18/Dto.kt
@@ -1,16 +1,0 @@
-package eu.kanade.tachiyomi.extension.vi.tranh18
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-class FilterOption(
-    val name: String,
-    val value: String,
-)
-
-@Serializable
-class FilterData(
-    val tags: List<FilterOption>,
-    val areas: List<FilterOption>,
-    val end: List<FilterOption>,
-)
