+++ src/vi/tranh18/src/eu/kanade/tachiyomi/extension/vi/tranh18/Filters.kt
@@ -1,13 +1,22 @@
 package eu.kanade.tachiyomi.extension.vi.tranh18
 
 import eu.kanade.tachiyomi.source.model.Filter
+import kotlinx.serialization.Serializable
 
-class Genre(val name: String, val genre: String) {
+@Serializable
+class FilterOption(val name: String, val value: String) {
     override fun toString() = name
 }
 
-class GenreList(genres: Array<Genre>) : Filter.Select<Genre>("Thể loại", genres)
+class GenreList(options: Array<FilterOption>) : Filter.Select<FilterOption>("Thể loại", options)
 
-class StatusList(status: Array<Genre>) : Filter.Select<Genre>("Tiến độ", status)
+class StatusList(status: Array<FilterOption>) : Filter.Select<FilterOption>("Tiến độ", status)
 
-class TagList(tags: Array<Genre>) : Filter.Select<Genre>("Từ khóa", tags)
+class TagList(tags: Array<FilterOption>) : Filter.Select<FilterOption>("Từ khóa", tags)
+
+@Serializable
+class FilterData(
+    val tags: List<FilterOption>,
+    val areas: List<FilterOption>,
+    val end: List<FilterOption>,
+)
